+++ app/routes/components/Select/changelog.codegen.ts
@@ -6,4 +6,5 @@ export const changelog = ''
   + '- 2025-09-25: Allow props to accept generics.\n'
   + "- 2025-09-25: Use `onToggle` instead of `onClick` in `Popover.Trigger`'s render\n"
   + '  prop.\n'
+  + '- 2025-10-03: Fix missing dependencies in useEffect hooks, etc.\n'
   ;