+++ src/components/Select/index.ts
@@ -1,3 +1,3 @@
-// Select from copy-ui @ 2025-09-25.
+// Select from copy-ui @ 2025-10-03.
 
 export { Select } from "./Select";