+++ src/components/Select/CHANGELOG.md
@@ -2,3 +2,4 @@
 - 2025-09-25: Allow props to accept generics.
 - 2025-09-25: Use `onToggle` instead of `onClick` in `Popover.Trigger`'s render
   prop.
+- 2025-10-03: Fix missing dependencies in useEffect hooks, etc.